+++ src/lib.rs
@@ -1,7 +1,9 @@
 pub mod hyperliquid;
 pub mod notion;
+pub mod redirect;
 pub mod structs;
 
 pub use hyperliquid::{HyperliquidMonitor, HyperliquidMonitorError};
 pub use notion::{NotionWriter, NotionWriterError};
+pub use redirect::{RedirectError, ResolvedSymbol, SymbolRedirects};
 pub use structs::{NotionRowData, PositionTradeEvent, TradeAction};